requirements.txt
get==0.0.20
post==0.0.12
public==0.0.37
PyQt5==5.7
pyqt5-tools==5.7.dev9
python-Levenshtein==0.12.0
query-string==0.0.11
request==0.0.12
requests==2.11.1
setupfiles==0.0.13
sip==4.18.1
